+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+05/06/2016 -> 1.13.x -> 1.14.0
+1. new API: /write/updateDomain
+2. Initial release supports update alias. More functionalities on the way.
+
 05/06/2016 -> 1.13.1 -> 1.13.2
 1. /read/getAccounts now returns domainId